Don't close LanguageRuntime in forkRun
The LanguageRuntime in forkRun is from a plugin host, which will cache
the runtimes it's asked for and close them when the host is closed.

Closing the runtime here leads to _at least_ an error in the logs when
the host tries to close it later and finds the process already gone, and
could potentially result in another part of the code asking for the same
runtime object and having it killed unexpectedly.